+++ src/index.tsx
@@ -1,9 +1,8 @@
 import * as React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import './index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
+import reportWebVitals from './test/reportWebVitals';
 
 import Store from './reducers/stores';
